use int32_t instead of long as a return type

+++ ThrEd4/thred.cpp
@@ -12892,11 +12892,11 @@ auto thred::getMoveDelta(const POINT& point) noexcept -> F_POINT {
   return F_POINT {wrap::toFloat(point.x) / HorizontalRatio, -wrap::toFloat(point.y) / VerticalRatio};
 }
 
-auto thred::scaleHorizontal(float const& value) noexcept -> long {
+auto thred::scaleHorizontal(float const& value) noexcept -> int32_t {
   return std::lround(value * HorizontalRatio);
 }
 
-auto thred::scaleVertical(float const& value) noexcept -> long {
+auto thred::scaleVertical(float const& value) noexcept -> int32_t {
   return std::lround(value * VerticalRatio);
 }
